Extract a subroutine that initializes rvp_trace_size_limit.  Use int64_t instead of ssize_t for the trace size and limit.  Statically initialize rvp_trace_size_limit to INT64_MAX so that there is no effective limit (well...) unless RVP_TRACE_SIZE_LIMIT is set in the environment.

+++ llvm/ngrt/trace.c
@@ -18,7 +18,7 @@
 #include "trace.h"
 #include "tracefmt.h"
 
-ssize_t rvp_trace_size = 0;
+int64_t rvp_trace_size = 0;
 
 static __section(".text") deltops_t deltops = { .matrix = { { 0 } } };
